requirements.txt
autopep8==1.5.1
imutils==0.5.3
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pyrealsense2==2.34.0.1470
tflite-runtime==2.1.0.post1
